write json graphs not just machine graph
